android: hover post button when narrow
