+++ crates/notedeck_columns/src/ui/mod.rs
@@ -6,6 +6,7 @@ pub mod configure_deck;
 pub mod edit_deck;
 pub mod images;
 pub mod note;
+pub mod post;
 pub mod preview;
 pub mod profile;
 pub mod relay;
